% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore StatFin databases}
\usage{
explore(db = NULL)
}
\arguments{
\item{db}{The id of a specific database (optional)}
}
\value{
A data frame containing the \code{id} and \code{name} of the databases
}
\description{
\code{explore} returns the IDs and names of StatFin databases
}
\details{
The function can be called without arguments or with the \code{db} argument.
Without arguments, it returns the StatFin main databases. If supplied a
specific database ID with the \code{db} argument, it returns the sub-databases
for that database.
}
