% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from Statistics Finland databases}
\usage{
get_data(db = NULL, code = NULL, filter = NULL, values = NULL, format = "df")
}
\arguments{
\item{db}{Code of database to be retrieved}

\item{code}{Variable(s) to use in filtering the results}

\item{filter}{Filtering type (see below)}

\item{values}{Values to filter for}

\item{format}{Format to return results in. Defaults to \code{df}}
}
\value{
Return data type depends on the value of \code{format}
}
\description{
The \code{get_data} function retrieves database data from Statistics Finland.
}
\details{
Define the database to be retrieved from using the argument \code{db}. If you are
unaware of your database's code, use the \code{\link[=explore]{explore()}} function.

The \code{code} argument defines which variables to filter by. Currently, it only
supports a single variable.

The \code{filter} argument can be one of five filter types:
\itemize{
\item \code{item}: Return the data for all filter values defined in \code{values}
\item \code{all}: Similar to \code{item}, but used for wildcard selection with the \code{*} symbol
\item \code{top}: Return the \code{values} top-most data
}

The values to be filtered by are defined in the \code{values} argument. Note,
that values \emph{must} be entered as character strings!

The argument \code{format} defines in which format your results are returned.

Currently supported formats are \code{df} (data frame) and \code{json} (JSON-formatted).
}
